Created some classes: - Repository classes - Service classes - PlaySong class
Have continued the work to be able to store a users preferred genres and what media the user has listened to. This is not done but a push is made to branch so that progress is visible and so that input can be highlighted.

Created User and also PreferredGenre class and PlayedSong class to begin with.

// Alfred

// Alfred